added more seed data and am beginning to create the wrightstream predictor

+++ db/migrations/20180627154918_purchases_supplies.js
@@ -2,7 +2,7 @@ exports.up = function(knex, Promise) {
   return knex.schema.createTable('purchases_supplies', (table) => {
     table.integer('purchase_id').nullable().references('purchases.id');
     table.integer('supplies_id').notNullable().references('supplies.id');
-    table.integer('supplies_qty').notNullable()
+    table.decimal('supplies_qty').notNullable()
     table.string('supplies_measurement').notNullable()
     table.boolean('completed').notNullable().defaultsTo(false);
     table.timestamps('true', 'true');
